FloatTextField updated Updated gem dependency AddTokenScene, AddTokenScene added "become first responder" on appear
